Minor clean architecture reorg
